FLOW-958 legends added
