+++ packages/flow-dashboard/src/components/f-timeseries-chart/f-timeseries-chart-global.scss
@@ -2,7 +2,9 @@
 f-timeseries-chart {
 	@include base();
 	display: flex;
+	flex-direction: column;
 	min-height: 200px;
+
 	svg {
 		.domain {
 			stroke: var(--color-border-secondary);
@@ -33,6 +35,7 @@ f-timeseries-chart {
 	.f-chart-tooltip {
 		position: fixed;
 		pointer-events: none;
+		z-index: 1;
 		&.hide {
 			display: none;
 		}
@@ -44,6 +47,25 @@ f-timeseries-chart {
 			top 0.2s linear,
 			bottom 0.2s linear,
 			right 0.2s linear;
+	}
+
+	.disable-legend {
+		opacity: 0.4;
+	}
+
+	.series-path.disable {
+		opacity: 0.4;
+		.bars {
+			fill-opacity: 0.3;
+		}
+	}
+	.series-path.active {
+		opacity: 1;
+		fill-opacity: 1 !important;
+		stroke-opacity: 1;
+		.bars {
+			fill-opacity: 1;
+		}
 	}
 }
